requirements.txt
blinker==1.9.0
click==8.3.1
flask==3.1.3
gevent==25.9.1
greenlet==3.3.1
gunicorn==25.1.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
packaging==26.0
werkzeug==3.1.6
zope-event==6.1
zope-interface==8.2
Flask-SQLAlchemy-Lite